+++ DesignLab/Define.cpp
@@ -8,6 +8,6 @@ const int Define::GOAL_TAPE = 1200;
 
 const bool Define::FLAG_GRAPHIC_AVAILABLE = true;
 
-const double Define::MY_PI = 3.141592653589793;
+const float Define::MY_PI = 3.141592653589793f;
 
 const bool Define::FLAG_DO_PRUNING = false;
