+++ DesignLab/HexapodConst.cpp
@@ -1,21 +1,21 @@
 #include "HexapodConst.h"
 
-const double HexapodConst::COXA_LENGTH = 52.0;
+const float HexapodConst::COXA_LENGTH = 52.0f;
 
-const double HexapodConst::FEMUR_LENGTH = 66.0;
+const float HexapodConst::FEMUR_LENGTH = 66.0f;
 
-const double HexapodConst::TIBIA_LENGTH = 130.0;
+const float HexapodConst::TIBIA_LENGTH = 130.0f;
 
-const double HexapodConst::BODY_FRONT_WIDTH = 60.0;
+const float HexapodConst::BODY_FRONT_WIDTH = 60.0f;
 
-const double HexapodConst::BODY_CENTER_WIDTH = 100.0;
+const float HexapodConst::BODY_CENTER_WIDTH = 100.0f;
 
-const double HexapodConst::BODY_REAR_WIDTH = 60.0;
+const float HexapodConst::BODY_REAR_WIDTH = 60.0f;
 
-const double HexapodConst::BODY_FRONT_LENGTH = 120;
+const float HexapodConst::BODY_FRONT_LENGTH = 120.0f;
 
-const double HexapodConst::BODY_REAR_LENGTH = 120;
+const float HexapodConst::BODY_REAR_LENGTH = 120.0f;
 
-const double HexapodConst::VERTICAL_MAX_RANGE = 190;
+const float HexapodConst::VERTICAL_MAX_RANGE = 190.0f;
 
-const double HexapodConst::VERTICAL_MIN_RANGE = 30;
+const float HexapodConst::VERTICAL_MIN_RANGE = 30.0f;
